requirements.txt
pytorch >= 1.9.0
torchvision
transformers
tqdm
pandas
timm
ftfy
regex
braceexpand
webdataset
jsonlines
tensorboard
